requirements.txt
mistralai
python-dotenv
termcolor
requests 